Ide updates, measuring qreg, runtime integration

+++ src/core/quantum/systems/ensemble.hpp
@@ -9,12 +9,24 @@
 namespace qlib {
 namespace quantum {
 
+/// <Summary>
+/// Represents a collection of multiple identical quantum systems
+/// </Summary>
 class ensemble : public qsystem{
     private:
+        /// <Summary>
+        /// List of all systems in this collection
+        /// </Summary>
         std::vector<qsystem*> set;
 
     public:
+        /// <Summary>
+        /// Create an ensemble from a given list of systems
+        /// </Summary>
         ensemble(std::vector<qsystem*> eset): set(eset){}
+        /// <Summary>
+        /// Delete the ensemble and cleanup pointers
+        /// </Summary>
         ~ensemble(){
             //Clean up memory
             for(std::vector<qsystem*>::iterator it = set.begin(); it != set.end(); it++) {
@@ -32,7 +44,7 @@ class ensemble : public qsystem{
         /// <Summary>
         /// Apply operator to all systems in the ensemble
         /// </Summary>
-        void apply(qlib::quantum::gates::igate& op,std::vector<ulong> inputBits){
+        void apply(qlib::quantum::gates::igate& op,std::vector<u64> inputBits){
             for(std::vector<qsystem*>::iterator it = set.begin(); it != set.end(); it++) {
                 (*it)->apply(op, inputBits);
             }
@@ -41,12 +53,12 @@ class ensemble : public qsystem{
         /// <Summary>
         /// Measure all systems in the ensemble and return most common answer
         /// </Summary>
-        i32 measure(i32 qubit){
-            std::map<i32, u32> map;
-            u32 max = 0;
+        i8 measure(i64 qubit){
+            std::map<i8, u64> map;
+            u64 max = 0;
             for(std::vector<qsystem*>::iterator it = set.begin(); it != set.end(); it++){
-                i32 r = (*it)->measure(qubit);
-                u32 count = ++map[r];
+                i8 r = (*it)->measure(qubit);
+                u64 count = ++map[r];
                 if(count > map[max]){
                     max = r;
                 }
